fix: extract resizeImage
